OrderStore: Move upsert logic to OrdersUpsertUseCase
